+++ lib/src/util/log.dart
@@ -1,0 +1,9 @@
+class Log {
+  Log._();
+
+  bool showLog = false;
+
+  static void log(Object msg) {
+    print(msg ?? "null");
+  }
+}